fix(configprovider): 调整demo展示
