--- src/packages/configprovider/demos/h5/demo4.tsx
@@ -1,15 +1,0 @@
-import React from 'react'
-import { ConfigProvider, TextArea } from '@nutui/nutui-react'
-import enUS from '@nutui/nutui-react/dist/es/locale/en-US'
-
-const Demo4 = () => {
-  return (
-    <>
-      <ConfigProvider locale={enUS}>
-        <TextArea disabled showCount maxLength={20} />
-      </ConfigProvider>
-    </>
-  )
-}
-
-export default Demo4
