--- src/packages/configprovider/demos/taro/demo4.tsx
@@ -1,15 +1,0 @@
-import React from 'react'
-import { ConfigProvider, TextArea } from '@nutui/nutui-react-taro'
-import enUS from '@nutui/nutui-react-taro/dist/es/locales/en-US'
-
-const Demo4 = () => {
-  return (
-    <>
-      <ConfigProvider locale={enUS}>
-        <TextArea disabled showCount maxLength={20} />
-      </ConfigProvider>
-    </>
-  )
-}
-
-export default Demo4
